Clock screen implemented!
I've got it just how I want it. I've gotten rid of most of the visual
glitches in the radial display, but there is one remaining glitch
but this is actually quite nice - it results in the display becoming
sparkly and I quite like it!

Next - implement the stats screen and implement the touch sensing