Modified display refresh conditions
Now a display refresh is allowed only when making the blocking call to the prime
checker. This was meant only to fix #11 (the guess was that we were getting a
refresh interrupt just as the red prime line had been drawn, painting it over
with a green line) but as a nice side-effect it seems to have dramatically
speeded up the clock, probbaly because we are now not doing a bunch of redundant
draws.

The histogram color scaling has been changed to fix #10 - the color is now scaled to
the highest bin value in a row.